Don't store a Twine across statements
It's never safe to keep a Twine instance around longer than the end of
the current statement.  So instead, turn it into a std::string before
the statement ends.

The LLVM documentation warns that "A Twine is not intended for use
directly and should not be stored, its implementation relies on the
ability to store pointers to temporary stack objects which may be
deallocated at the end of a statement."